+++ .gitignore
@@ -3,4 +3,3 @@ lib/
 types/
 node_modules/
 .vscode/
-coverage/
