+++ src/interpreter/index.ts
@@ -41,7 +41,7 @@ import { isBoxable, isUnboxable } from "../brsTypes/Boxing";
 
 import { ComponentDefinition } from "../componentprocessor";
 import pSettle from "p-settle";
-import { CoverageCollector } from "../coverageLib";
+import { CoverageCollector } from "../coverage";
 
 /** The set of options used to configure an interpreter's execution. */
 export interface ExecutionOptions {
